UIEH-1194 Move component to a separate folder
